[LAND-117]Rewrite Image to TypeScript
